Minor Improvements to Multiplayers, A lot of new Commenting.
BigBossServer:
I have done some minor restructuring in the Multiplayer Server side of the program. A lot of Commenting has also been done. The Support of Casted Abilities has also been added to the Server.

Bigboss:
No restructuring has been done, A lot of commenting has been added to some aspects to the Code. Over the next few updates the I will continue to add more comments to make it easier for the end-user to create new Abilities.

Abilities:
AbilityBase & AbilityCarcus: Will be refactored in the next few updates, The Flash and heal Abilities will be Updated to work with the new and improved base and Carcus *if they still exist*.
Flash: In the next update the The Flash Ability is going to get Animation on both the start point and the final point.
Heal: Will Have to be redone, *at this point its too confusing for me to fix its Glitches (Instead of Healing the player it spawns a new Entity in to the world, and after enough spamming it heals the players).